backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-genai==1.14.0
python-dotenv==1.0.1
websockets==14.2
Pillow==11.1.0
groq>=0.4.2
edge-tts>=7.0.0
